11 - Agentes com CrewAI e Streamlit/Aula/How_to_implement_a_simple_Ui/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
crewai==0.22.5
dataclasses-json==0.6.7
deprecated==1.2.18
distro==1.9.0
docstring-parser==0.15
exceptiongroup==1.2.2
frozenlist==1.6.0
gitdb==4.0.12
gitpython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
instructor==0.5.2
jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.0.5
langchain-text-splitters==0.0.2
langsmith==0.1.147
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.4.3
mypy-extensions==1.1.0
narwhals==1.37.0
numpy==1.26.4
openai==1.76.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.16
packaging==23.2
pandas==2.2.3
pillow==10.4.0
propcache==0.3.1
protobuf==4.25.7
pyarrow==20.0.0
pydantic==2.11.3
pydantic-core==2.33.1
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2023.12.25
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.40
streamlit==1.32.2
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.9.4
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
